fix https bitbucket url regex of the form https://[username]@bitbucket.org/[username]/[repo name].git -- missing the "[username]@" part
`:Gbrowse` wasn't working for my bitbucket https urls. I kept getting `fugitive: "origin' is not a supported remote`.
I realised that the regex on `line 15` was missing the 'username@' part of 'https://username@bitbucket.org...'. 
Added `\%(.\{-\}@\)\=` and now it works. Not 100% sure if this is the best regex to do this and hopefully it won't mess up anything else but I can now `:Gbrowse` with my bitbucket project.